vision/requirements.txt
# Seedy Vision — Dependencies
# pip install -r requirements.txt

# ── Core ML ──────────────────────────────────────────
ultralytics>=8.3.0           # YOLOv8/v11 training & inference
torch>=2.2.0
torchvision>=0.17.0
onnx>=1.16.0
onnxruntime-gpu>=1.17.0

# ── Dataset tools ────────────────────────────────────
roboflow>=1.1.0              # Roboflow dataset API
kaggle>=1.6.0                # Kaggle dataset downloads
datasets>=2.19.0             # HuggingFace datasets
fiftyone>=0.24.0             # Dataset visualization & curation

# ── Image processing ────────────────────────────────
opencv-python-headless>=4.9.0
Pillow>=10.3.0
albumentations>=1.4.0        # Data augmentation
imagededup>=0.3.2            # Duplicate image detection

# ── Data & config ───────────────────────────────────
pyyaml>=6.0
pandas>=2.2.0
tqdm>=4.66.0
requests>=2.31.0
httpx>=0.27.0

# ── Tracking & logging ──────────────────────────────
mlflow>=2.12.0               # Experiment tracking (optional)
tensorboard>=2.16.0

# ── Utilities ────────────────────────────────────────
rich>=13.7.0                 # Pretty terminal output
typer>=0.12.0                # CLI framework
